src/4.advanced_opengl/10.2.asteroids/requirements.txt
PyOpenGL==3.1.5
glfw==2.1.0
PyGLM==2.1.0
Pillow==8.2.0
assimp-py